+++ examples/remote-peering-connections/rpc-requestor/variables.tf
@@ -11,5 +11,6 @@ variable "region" {}
 
 variable "network_configuration" {
   type = any
+  default = null
 }