+++ RELEASE-NOTES.md
@@ -1,10 +1,16 @@
+# April 16, 2024 Release Notes - 0.6.6
+
+## Updates
+1. Module now supports external dependencies on private IP addresses, DRG attachments, remote peering connections and local peering gateways. See [External Dependencies](./README.md#ext-dep) for details.
+2. All dependency variables are now strongly typed, enhancing usage guidance.
+
 # April 08, 2024 Release Notes - 0.6.5
 ## Additions
 1. MVP module for Network Load Balancers.
 
 ## Updates
 1. Module dependency on externally managed network resources enhanced, including improved examples and documentation.
-3. Release tracking via freeform tags.
+2. Release tracking via freeform tags.
 
 ## Fixes
 1. L7 load balancers module dependency on compartments.